+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.3.4] - 2026-09-11
+
+### Changed
+
+- Updated dependencies.
+
 ## [0.3.3] - 2026-09-04
 
 ### Added