+++ packages/server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.14] - 2026-09-11
+
+### Changed
+
+- Updated dependencies.
+
 ## [0.1.13] - 2026-09-09
 
 ### Changed